+++ Libplanet.Tests/Store/FileStoreTest.cs
@@ -26,13 +26,14 @@ namespace Libplanet.Tests.Store
         {
             Assert.Empty(_fx.Store.ListNamespaces());
 
-            _fx.Store.PutBlock(_ns, _fx.Block1);
+            _fx.Store.PutBlock(_fx.Block1);
+            _fx.Store.AppendIndex(_ns, _fx.Block1.Hash);
             Assert.Equal(
                 new[] { _ns }.ToImmutableHashSet(),
                 _fx.Store.ListNamespaces().ToImmutableHashSet()
             );
 
-            _fx.Store.PutBlock("asdf", _fx.Block1);
+            _fx.Store.AppendIndex("asdf", _fx.Block1.Hash);
             Assert.Equal(
                 new[] { _ns, "asdf" }.ToImmutableHashSet(),
                 _fx.Store.ListNamespaces().ToImmutableHashSet()
@@ -43,8 +44,8 @@ namespace Libplanet.Tests.Store
         public void CanReturnTransactionPath()
         {
             Assert.Equal(
-                Path.Combine(_fx.Path, _ns, "tx", "45a2", "2187e2d8850bb357886958bc3e8560929ccc886958bc3e8560929ccc9ccc"),
-                _fx.Store.GetTransactionPath(_ns, _fx.TxId1)
+                Path.Combine(_fx.Path, "tx", "45a2", "2187e2d8850bb357886958bc3e8560929ccc886958bc3e8560929ccc9ccc"),
+                _fx.Store.GetTransactionPath(_fx.TxId1)
             );
         }
 
@@ -52,31 +53,22 @@ namespace Libplanet.Tests.Store
         public void CanReturnBlockPath()
         {
             Assert.Equal(
-                Path.Combine(_fx.Path, _ns, "blocks", "45a2", "2187e2d8850bb357886958bc3e8560929ccc886958bc3e8560929ccc9ccc"),
-                _fx.Store.GetBlockPath(_ns, _fx.Hash1)
+                Path.Combine(_fx.Path, "blocks", "45a2", "2187e2d8850bb357886958bc3e8560929ccc886958bc3e8560929ccc9ccc"),
+                _fx.Store.GetBlockPath(_fx.Hash1)
             );
         }
 
         [Fact]
-        public void CanReturnStagedTransactionPath()
+        public void ReturnStagedTransactionPath()
         {
             Assert.Equal(
-                Path.Combine(_fx.Path, _ns, "stage", "45a22187e2d8850bb357886958bc3e8560929ccc886958bc3e8560929ccc9ccc"),
-                _fx.Store.GetStagedTransactionPath(_ns, _fx.TxId1)
+                Path.Combine(_fx.Path, "stage", "45a22187e2d8850bb357886958bc3e8560929ccc886958bc3e8560929ccc9ccc"),
+                _fx.Store.GetStagedTransactionPath(_fx.TxId1)
             );
         }
 
         [Fact]
-        public void CanReturnAddressPath()
-        {
-            Assert.Equal(
-                Path.Combine(_fx.Path, _ns, "addr", "45a2", "2187e2D8850bb357886958bC3E8560929ccc"),
-                _fx.Store.GetAddressPath(_ns, _fx.Address1)
-            );
-        }
-
-        [Fact]
-        public void CanReturnStatesPath()
+        public void ReturnStatesPath()
         {
             var hash = new HashDigest<SHA256>(new byte[]
             {
@@ -86,134 +78,134 @@ namespace Libplanet.Tests.Store
                 0x9c, 0xcc,
             });
             Assert.Equal(
-                Path.Combine(_fx.Path, _ns, "states", "45a2", "2187e2d8850bb357886958bc3e8560929ccc886958bc3e8560929ccc9ccc"),
-                _fx.Store.GetStatesPath(_ns, hash)
+                Path.Combine(_fx.Path, "states", "45a2", "2187e2d8850bb357886958bc3e8560929ccc886958bc3e8560929ccc9ccc"),
+                _fx.Store.GetStatesPath(hash)
             );
         }
 
         [Fact]
-        public void CanStoreBlock()
+        public void StoreBlock()
         {
-            Assert.Empty(_fx.Store.IterateBlockHashes(_ns));
-            Assert.Null(_fx.Store.GetBlock<BaseAction>(_ns, _fx.Block1.Hash));
-            Assert.Null(_fx.Store.GetBlock<BaseAction>(_ns, _fx.Block2.Hash));
-            Assert.Null(_fx.Store.GetBlock<BaseAction>(_ns, _fx.Block3.Hash));
-            Assert.False(_fx.Store.DeleteBlock(_ns, _fx.Block1.Hash));
+            Assert.Empty(_fx.Store.IterateBlockHashes());
+            Assert.Null(_fx.Store.GetBlock<BaseAction>(_fx.Block1.Hash));
+            Assert.Null(_fx.Store.GetBlock<BaseAction>(_fx.Block2.Hash));
+            Assert.Null(_fx.Store.GetBlock<BaseAction>(_fx.Block3.Hash));
+            Assert.False(_fx.Store.DeleteBlock(_fx.Block1.Hash));
 
-            _fx.Store.PutBlock(_ns, _fx.Block1);
-            Assert.Equal(1, _fx.Store.CountBlocks(_ns));
+            _fx.Store.PutBlock(_fx.Block1);
+            Assert.Equal(1, _fx.Store.CountBlocks());
             Assert.Equal(
                 new HashSet<HashDigest<SHA256>>
                 {
                     _fx.Block1.Hash,
                 },
-                _fx.Store.IterateBlockHashes(_ns).ToHashSet());
+                _fx.Store.IterateBlockHashes().ToHashSet());
             Assert.Equal(
                 _fx.Block1,
-                _fx.Store.GetBlock<BaseAction>(_ns, _fx.Block1.Hash));
-            Assert.Null(_fx.Store.GetBlock<BaseAction>(_ns, _fx.Block2.Hash));
-            Assert.Null(_fx.Store.GetBlock<BaseAction>(_ns, _fx.Block3.Hash));
+                _fx.Store.GetBlock<BaseAction>(_fx.Block1.Hash));
+            Assert.Null(_fx.Store.GetBlock<BaseAction>(_fx.Block2.Hash));
+            Assert.Null(_fx.Store.GetBlock<BaseAction>(_fx.Block3.Hash));
 
-            _fx.Store.PutBlock(_ns, _fx.Block2);
-            Assert.Equal(2, _fx.Store.CountBlocks(_ns));
+            _fx.Store.PutBlock(_fx.Block2);
+            Assert.Equal(2, _fx.Store.CountBlocks());
             Assert.Equal(
                 new HashSet<HashDigest<SHA256>>
                 {
                     _fx.Block1.Hash,
                     _fx.Block2.Hash,
                 },
-                _fx.Store.IterateBlockHashes(_ns).ToHashSet());
+                _fx.Store.IterateBlockHashes().ToHashSet());
             Assert.Equal(
                 _fx.Block1,
-                _fx.Store.GetBlock<BaseAction>(_ns, _fx.Block1.Hash));
+                _fx.Store.GetBlock<BaseAction>(_fx.Block1.Hash));
             Assert.Equal(
                 _fx.Block2,
-                _fx.Store.GetBlock<BaseAction>(_ns, _fx.Block2.Hash));
-            Assert.Null(_fx.Store.GetBlock<BaseAction>(_ns, _fx.Block3.Hash));
+                _fx.Store.GetBlock<BaseAction>(_fx.Block2.Hash));
+            Assert.Null(_fx.Store.GetBlock<BaseAction>(_fx.Block3.Hash));
 
-            Assert.True(_fx.Store.DeleteBlock(_ns, _fx.Block1.Hash));
-            Assert.Equal(1, _fx.Store.CountBlocks(_ns));
+            Assert.True(_fx.Store.DeleteBlock(_fx.Block1.Hash));
+            Assert.Equal(1, _fx.Store.CountBlocks());
             Assert.Equal(
                 new HashSet<HashDigest<SHA256>>
                 {
                     _fx.Block2.Hash,
                 },
-                _fx.Store.IterateBlockHashes(_ns).ToHashSet());
-            Assert.Null(_fx.Store.GetBlock<BaseAction>(_ns, _fx.Block1.Hash));
+                _fx.Store.IterateBlockHashes().ToHashSet());
+            Assert.Null(_fx.Store.GetBlock<BaseAction>(_fx.Block1.Hash));
             Assert.Equal(
                 _fx.Block2,
-                _fx.Store.GetBlock<BaseAction>(_ns, _fx.Block2.Hash));
-            Assert.Null(_fx.Store.GetBlock<BaseAction>(_ns, _fx.Block3.Hash));
+                _fx.Store.GetBlock<BaseAction>(_fx.Block2.Hash));
+            Assert.Null(_fx.Store.GetBlock<BaseAction>(_fx.Block3.Hash));
         }
 
         [Fact]
-        public void CanStoreTx()
+        public void StoreTx()
         {
-            Assert.Equal(0, _fx.Store.CountTransactions(_ns));
-            Assert.Empty(_fx.Store.IterateTransactionIds(_ns));
-            Assert.Null(_fx.Store.GetTransaction<BaseAction>(_ns, _fx.Transaction1.Id));
-            Assert.Null(_fx.Store.GetTransaction<BaseAction>(_ns, _fx.Transaction2.Id));
-            Assert.False(_fx.Store.DeleteTransaction(_ns, _fx.Transaction1.Id));
+            Assert.Equal(0, _fx.Store.CountTransactions());
+            Assert.Empty(_fx.Store.IterateTransactionIds());
+            Assert.Null(_fx.Store.GetTransaction<BaseAction>(_fx.Transaction1.Id));
+            Assert.Null(_fx.Store.GetTransaction<BaseAction>(_fx.Transaction2.Id));
+            Assert.False(_fx.Store.DeleteTransaction(_fx.Transaction1.Id));
 
-            _fx.Store.PutTransaction(_ns, _fx.Transaction1);
-            Assert.Equal(1, _fx.Store.CountTransactions(_ns));
+            _fx.Store.PutTransaction(_fx.Transaction1);
+            Assert.Equal(1, _fx.Store.CountTransactions());
             Assert.Equal(
                 new HashSet<TxId>
                 {
                     _fx.Transaction1.Id,
                 },
-                _fx.Store.IterateTransactionIds(_ns)
+                _fx.Store.IterateTransactionIds()
             );
             Assert.Equal(
                 _fx.Transaction1,
-                _fx.Store.GetTransaction<BaseAction>(_ns, _fx.Transaction1.Id)
+                _fx.Store.GetTransaction<BaseAction>(_fx.Transaction1.Id)
             );
-            Assert.Null(_fx.Store.GetTransaction<BaseAction>(_ns, _fx.Transaction2.Id));
+            Assert.Null(_fx.Store.GetTransaction<BaseAction>(_fx.Transaction2.Id));
 
-            _fx.Store.PutTransaction(_ns, _fx.Transaction2);
-            Assert.Equal(2, _fx.Store.CountTransactions(_ns));
+            _fx.Store.PutTransaction(_fx.Transaction2);
+            Assert.Equal(2, _fx.Store.CountTransactions());
             Assert.Equal(
                 new HashSet<TxId>
                 {
                     _fx.Transaction1.Id,
                     _fx.Transaction2.Id,
                 },
-                _fx.Store.IterateTransactionIds(_ns).ToHashSet()
+                _fx.Store.IterateTransactionIds().ToHashSet()
             );
             Assert.Equal(
                 _fx.Transaction1,
-                _fx.Store.GetTransaction<BaseAction>(_ns, _fx.Transaction1.Id)
+                _fx.Store.GetTransaction<BaseAction>(_fx.Transaction1.Id)
             );
             Assert.Equal(
                 _fx.Transaction2,
-                _fx.Store.GetTransaction<BaseAction>(_ns, _fx.Transaction2.Id));
+                _fx.Store.GetTransaction<BaseAction>(_fx.Transaction2.Id));
 
-            Assert.True(_fx.Store.DeleteTransaction(_ns, _fx.Transaction1.Id));
-            Assert.Equal(1, _fx.Store.CountTransactions(_ns));
+            Assert.True(_fx.Store.DeleteTransaction(_fx.Transaction1.Id));
+            Assert.Equal(1, _fx.Store.CountTransactions());
             Assert.Equal(
                 new HashSet<TxId>
                 {
                     _fx.Transaction2.Id,
                 },
-                _fx.Store.IterateTransactionIds(_ns)
+                _fx.Store.IterateTransactionIds()
             );
-            Assert.Null(_fx.Store.GetTransaction<BaseAction>(_ns, _fx.Transaction1.Id));
+            Assert.Null(_fx.Store.GetTransaction<BaseAction>(_fx.Transaction1.Id));
             Assert.Equal(
                 _fx.Transaction2,
-                _fx.Store.GetTransaction<BaseAction>(_ns, _fx.Transaction2.Id)
+                _fx.Store.GetTransaction<BaseAction>(_fx.Transaction2.Id)
             );
         }
 
         [Fact]
-        public void CanStoreIndex()
+        public void StoreIndex()
         {
-            Assert.Equal(0u, _fx.Store.CountIndex(_ns));
+            Assert.Equal(0, _fx.Store.CountIndex(_ns));
             Assert.Empty(_fx.Store.IterateIndex(_ns));
             Assert.Null(_fx.Store.IndexBlockHash(_ns, 0));
             Assert.Null(_fx.Store.IndexBlockHash(_ns, -1));
 
             Assert.Equal(0, _fx.Store.AppendIndex(_ns, _fx.Hash1));
-            Assert.Equal(1u, _fx.Store.CountIndex(_ns));
+            Assert.Equal(1, _fx.Store.CountIndex(_ns));
             Assert.Equal(
                 new List<HashDigest<SHA256>>()
                 {
@@ -224,7 +216,7 @@ namespace Libplanet.Tests.Store
             Assert.Equal(_fx.Hash1, _fx.Store.IndexBlockHash(_ns, -1));
 
             Assert.Equal(1, _fx.Store.AppendIndex(_ns, _fx.Hash2));
-            Assert.Equal(2u, _fx.Store.CountIndex(_ns));
+            Assert.Equal(2, _fx.Store.CountIndex(_ns));
             Assert.Equal(
                 new List<HashDigest<SHA256>>()
                 {
@@ -239,14 +231,13 @@ namespace Libplanet.Tests.Store
         }
 
         [Fact]
-        public void CanStoreStage()
+        public void StoreStage()
         {
-            _fx.Store.PutTransaction(_ns, _fx.Transaction1);
-            _fx.Store.PutTransaction(_ns, _fx.Transaction2);
-            Assert.Empty(_fx.Store.IterateStagedTransactionIds(_ns));
+            _fx.Store.PutTransaction(_fx.Transaction1);
+            _fx.Store.PutTransaction(_fx.Transaction2);
+            Assert.Empty(_fx.Store.IterateStagedTransactionIds());
 
             _fx.Store.StageTransactionIds(
-                _ns,
                 new HashSet<TxId>()
                 {
                     _fx.Transaction1.Id,
@@ -258,10 +249,9 @@ namespace Libplanet.Tests.Store
                     _fx.Transaction1.Id,
                     _fx.Transaction2.Id,
                 },
-                _fx.Store.IterateStagedTransactionIds(_ns).ToHashSet());
+                _fx.Store.IterateStagedTransactionIds().ToHashSet());
 
             _fx.Store.UnstageTransactionIds(
-                _ns,
                 new HashSet<TxId>
                 {
                     _fx.Transaction1.Id,
@@ -271,63 +261,13 @@ namespace Libplanet.Tests.Store
                 {
                     _fx.Transaction2.Id,
                 },
-                _fx.Store.IterateStagedTransactionIds(_ns).ToHashSet());
+                _fx.Store.IterateStagedTransactionIds().ToHashSet());
         }
 
         [Fact]
-        public void CanStoreAddress()
+        public void StoreBlockState()
         {
-            Assert.Equal(0, _fx.Store.CountAddresses(_ns));
-            Assert.Empty(_fx.Store.IterateAddresses(_ns));
-            Assert.Empty(_fx.Store.GetAddressTransactionIds(_ns, _fx.Address1));
-            Assert.Empty(_fx.Store.GetAddressTransactionIds(_ns, _fx.Address2));
-
-            // Add TxId1 to Address1
-            Assert.Equal(
-                0,
-                _fx.Store.AppendAddressTransactionId(_ns, _fx.Address1, _fx.TxId1));
-            Assert.Equal(1, _fx.Store.CountAddresses(_ns));
-            Assert.Equal(
-                new List<Address>() { _fx.Address1 },
-                _fx.Store.IterateAddresses(_ns));
-            Assert.Equal(
-                new List<TxId>() { _fx.TxId1 },
-                _fx.Store.GetAddressTransactionIds(_ns, _fx.Address1));
-            Assert.Empty(_fx.Store.GetAddressTransactionIds(_ns, _fx.Address2));
-
-            // Add TxId2 to Address1
-            Assert.Equal(
-                1,
-                _fx.Store.AppendAddressTransactionId(_ns, _fx.Address1, _fx.TxId2));
-            Assert.Equal(1, _fx.Store.CountAddresses(_ns));
-            Assert.Equal(
-                new List<Address>() { _fx.Address1 },
-                _fx.Store.IterateAddresses(_ns));
-            Assert.Equal(
-                new List<TxId>() { _fx.TxId1, _fx.TxId2 },
-                _fx.Store.GetAddressTransactionIds(_ns, _fx.Address1));
-            Assert.Empty(_fx.Store.GetAddressTransactionIds(_ns, _fx.Address2));
-
-            // Add TxId3 to Address2
-            Assert.Equal(
-                0,
-                _fx.Store.AppendAddressTransactionId(_ns, _fx.Address2, _fx.TxId3));
-            Assert.Equal(2, _fx.Store.CountAddresses(_ns));
-            Assert.Equal(
-                new List<Address>() { _fx.Address1, _fx.Address2 },
-                _fx.Store.IterateAddresses(_ns));
-            Assert.Equal(
-                new List<TxId>() { _fx.TxId1, _fx.TxId2 },
-                _fx.Store.GetAddressTransactionIds(_ns, _fx.Address1));
-            Assert.Equal(
-                new List<TxId>() { _fx.TxId3 },
-                _fx.Store.GetAddressTransactionIds(_ns, _fx.Address2));
-        }
-
-        [Fact]
-        public void CanStoreBlockState()
-        {
-            Assert.Empty(_fx.Store.GetBlockStates(_ns, _fx.Hash1));
+            Assert.Empty(_fx.Store.GetBlockStates(_fx.Hash1));
             AddressStateMap states = new AddressStateMap(
                 new Dictionary<Address, object>()
                 {
@@ -341,9 +281,9 @@ namespace Libplanet.Tests.Store
                     },
                 }.ToImmutableDictionary()
             );
-            _fx.Store.SetBlockStates(_ns, _fx.Hash1, states);
+            _fx.Store.SetBlockStates(_fx.Hash1, states);
 
-            AddressStateMap actual = _fx.Store.GetBlockStates(_ns, _fx.Hash1);
+            AddressStateMap actual = _fx.Store.GetBlockStates(_fx.Hash1);
             Assert.Equal(states[_fx.Address1], actual[_fx.Address1]);
             Assert.Equal(states[_fx.Address2], actual[_fx.Address2]);
         }