+++ Libplanet/Store/BaseStore.cs
@@ -27,98 +27,53 @@ namespace Libplanet.Store
             HashDigest<SHA256> hash
         );
 
-        public abstract IEnumerable<Address> IterateAddresses(
-            string @namespace
-        );
-
-        public abstract IEnumerable<TxId> GetAddressTransactionIds(
-            string @namespace,
-            Address address
-        );
-
-        public abstract long AppendAddressTransactionId(
-            string @namespace,
-            Address address,
-            TxId txId
-        );
-
         public abstract void StageTransactionIds(
-            string @namespace,
             ISet<TxId> txids
         );
 
         public abstract void UnstageTransactionIds(
-            string @namespace,
             ISet<TxId> txids
         );
 
-        public abstract IEnumerable<TxId> IterateStagedTransactionIds(
-            string @namespace
-        );
+        public abstract IEnumerable<TxId> IterateStagedTransactionIds();
 
-        public abstract IEnumerable<TxId> IterateTransactionIds(
-            string @namespace
-        );
+        public abstract IEnumerable<TxId> IterateTransactionIds();
 
-        public abstract Transaction<T> GetTransaction<T>(
-            string @namespace,
-            TxId txid
-        )
+        public abstract Transaction<T> GetTransaction<T>(TxId txid)
             where T : IAction;
 
-        public abstract void PutTransaction<T>(
-            string @namespace,
-            Transaction<T> tx
-        )
+        public abstract void PutTransaction<T>(Transaction<T> tx)
             where T : IAction;
 
-        public abstract bool DeleteTransaction(string @namespace, TxId txid);
+        public abstract bool DeleteTransaction(TxId txid);
 
-        public abstract IEnumerable<HashDigest<SHA256>> IterateBlockHashes(
-            string @namespace
-        );
+        public abstract IEnumerable<HashDigest<SHA256>> IterateBlockHashes();
 
-        public abstract Block<T> GetBlock<T>(
-            string @namespace,
-            HashDigest<SHA256> blockHash
-        )
+        public abstract Block<T> GetBlock<T>(HashDigest<SHA256> blockHash)
             where T : IAction;
 
-        public abstract void PutBlock<T>(
-            string @namespace,
-            Block<T> block
-        )
+        public abstract void PutBlock<T>(Block<T> block)
             where T : IAction;
 
-        public abstract bool DeleteBlock(
-            string @namespace,
-            HashDigest<SHA256> blockHash
-        );
+        public abstract bool DeleteBlock(HashDigest<SHA256> blockHash);
 
         public abstract AddressStateMap GetBlockStates(
-            string @namespace,
             HashDigest<SHA256> blockHash
         );
 
         public abstract void SetBlockStates(
-            string @namespace,
             HashDigest<SHA256> blockHash,
             AddressStateMap states
         );
 
-        public int CountTransactions(string @namespace)
+        public long CountTransactions()
         {
-            return IterateTransactionIds(@namespace).Count();
+            return IterateTransactionIds().LongCount();
         }
 
-        public int CountBlocks(string @namespace)
+        public long CountBlocks()
         {
-            return IterateBlockHashes(@namespace).Count();
-        }
-
-        public int CountAddresses(string @namespace)
-        {
-            return IterateAddresses(@namespace).Count();
+            return IterateBlockHashes().LongCount();
         }
 
         public abstract bool DeleteIndex(